+++ ChangeLog.md
@@ -16,6 +16,7 @@
     - Fix handling of Accounts URLs on..Link even after application startup
     - Make sure dropdown menu items are reactive to status changes
     - Close the modal after successful password change and validate reset password link
+    - Display back an indicator of the strength of new passwords
 
 ### 2.1.1